Improve ping handling (#22)
As suggested in this issue thread, there needs to be a way for
clients that send infrequent messages to make sure that the
connection is still open. Ping/pong are the natural way to do
this, but the ping behavior needs to be a bit more robust.

This commit adds a `wait_pong()` method that waits for a pong
to arrive and returns its payload.