Implement ping/pong matching (#22)
As suggested in the discussion thread, this commit mimics the
ping/pong API in aaugstin's websockets. A call to ping waits for
a response with the same payload. (An exception is raised if the
payload matches a ping that's already in flight.) A payload is
randomly generated if omitted by the caller. Add a new pong()
method that can be used for sending an unsolicited pong.