Allow more control over handshake (#4)
First attempt at this. Server handlers now receive a WebSocketRequest
instead of a WebSocketConnection. The request object provides the
client's handshake parameters and allows the server to finalize
handshake parameters. The server calls the requests ``accept()``
method to finish the handshake and obtain a connection object.

This first version exposes the request URL, headers, and subprotocols
to the server. The server is allowed to select a subprotocol. In future
versions (given support in wsproto), we can expand the request object
to customize headers and choose extensions.

In addition, the tests are updated: one new test to cover the
selection of a subprotocol, plus updates to existing tests to handle
the new request->connection step.

Finally, the README and example server are updated to show the new
request->connection step.