Delay connection closed (#69)
As described in the issue, get_message() was raising connection closed
even if there were pending messages. Per Nathaniel's suggestion, the
proper behavior is this:

1. If the remote endpoint closed the connection, then the local endpoint
   may continue reading all messages sent prior to closing.
2. If the local endpoint closed the connection, then the local endpoint
   may not read any more messages.

I added tests for these two conditions and implemented the behavior by
closing the ReceiveChannel inside the connection's `aclose()`. This
requires a bit of additional exception handling inside `get_message()`
and inside the reader task.

One slight surprise is that the test can't be written due to the bug
in #74! The client would hang because the reader task is blocked by
the unconsumed messages. So I changed the channel size to 32, which
allows this test to work, and I will replace this hard-coded value
when I fix #74.