Change send_message() behavior when close handshake is in progress
send_message() is changed to raise ConnectionClosed when a close
handshake is in progress.  Previously it would silently ignore
the call, which was an oversight, given that ConnectionClosed is
defined to cover connections "closed or in the process of closing".

Significantly, this fixes send_message() leaking a wsproto exception
(LocalProtocolError) with wsproto >= 1.22.0.

Fixes #175.