Use correct port for wss URLs when ssl_context is None (#197)
Prior to
https://github.com/python-trio/trio-websocket/commit/49b93c1b5a7977b5e37b6b16650b6a41c2064f98,
`_url_to_host()` modified `ssl_context` according to whether the URL
contained `ws` or `wss`. That commit introduced `return_ssl_context`
because the return value has a different type set (`ssl.SSLContext |
bool`) than the passed in argument (`ssl.SSLContext | None`). But the
determination of which port number to use was still based on
`ssl_context`, which would result in port 80 being used in the case
where the user passed in a `wss://` URL without an SSLContext. Checking
`return_ssl_context` instead results in the correct behavior.